+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 1.0.0 (Unreleased)
 
+IMPROVEMENTS:
+
+- add example of java autoscaling group [GH-2]( https://github.com/terraform-alicloud-modules/terraform-alicloud-market-app-on-autoscaling/pull/2)
+
 - **NEW:**  `market-app-on-autoscaling` [GH-1]( https://github.com/terraform-alicloud-modules/terraform-alicloud-market-app-on-autoscaling/pull/1)